requirements-TF-2.txt
tensorflow-gpu
tensorflow_probability
